Move bughunt native adapters into module
